improved reporting details , TransactionDetailsResponse class creation , Add new scenarios to the feature file, Modified existing class files to accomodate the new code

+++ src/main/java/pojo/AccountRequest.java
@@ -1,13 +1,13 @@
 package pojo;
 
 public class AccountRequest {
-    public int accountNumber;
+    public long accountNumber;
     public BankInformation bankInformation;
     public String accountStatus, accountType;
     public double accountBalance;
     public String accountCreated;
 
-    public AccountRequest(int accNum, BankInformation bankInfo, String status,
+    public AccountRequest(long accNum, BankInformation bankInfo, String status,
                           String type, double balance, String created) {
         this.accountNumber = accNum;
         this.bankInformation = bankInfo;
