+++ src/main/java/pojo/TransferRequest.java
@@ -1,11 +1,11 @@
 package pojo;
 
 public class TransferRequest {
-    public int fromAccountNumber;
-    public int toAccountNumber;
+    public long fromAccountNumber;
+    public long toAccountNumber;
     public double transferAmount;
 
-    public TransferRequest(int from, int to, double amount) {
+    public TransferRequest(long from, long to, double amount) {
         this.fromAccountNumber = from;
         this.toAccountNumber = to;
         this.transferAmount = amount;
